[TAS-401]fix: 인증 관련 초기화 로직 수정
