+++ App.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { Alert, StyleSheet, View } from 'react-native';
+import { Alert, AppState, StyleSheet, View } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { PaperProvider } from 'react-native-paper';
@@ -16,6 +16,7 @@ import { Signup } from 'screens/Signup';
 import { useAuthStore } from 'stores/auth';
 import { DialogProvider } from 'components/common/Dialog/Provider';
 import { LoadingOverlay } from 'components/common/LoadingOverlay';
+import { useRefreshTokenQuery } from 'hooks/queries/auth/useRefreshTokenQuery';
 
 const queryClient = new QueryClient();
 // 전역 에러 핸들링(모든 쿼리/뮤테이션)
@@ -36,18 +37,102 @@ queryClient.getMutationCache().subscribe(event => {
 });
 
 function AppContent() {
+  const {
+    tokenInfo,
+    isNeedRefreshToken,
+    isLoggedIn,
+    isNeedSignUp,
+    isHydrated,
+    setTokenInfo,
+    setIsLoggedIn,
+    setIsNeedRefreshToken,
+    setIsNeedSignUp,
+    removeTokenInfo,
+  } = useAuthStore(
+    ({
+      tokenInfo,
+      isNeedRefreshToken,
+      isLoggedIn,
+      isNeedSignUp,
+      isHydrated,
+      actions: { setTokenInfo, setIsLoggedIn, setIsNeedRefreshToken, setIsNeedSignUp, removeTokenInfo },
+    }) => ({
+      tokenInfo,
+      isNeedRefreshToken,
+      isLoggedIn,
+      isNeedSignUp,
+      isHydrated,
+      setTokenInfo,
+      setIsLoggedIn,
+      setIsNeedRefreshToken,
+      setIsNeedSignUp,
+      removeTokenInfo,
+    }),
+  );
+
   const isFetching = useIsFetching();
   const isMutating = useIsMutating();
   const isLoading = isFetching + isMutating > 0;
 
-  const { isLoggedIn, isNeedSignUp, isHydrated, removeTokenInfo } = useAuthStore(
-    ({ isLoggedIn, isNeedSignUp, isHydrated, actions: { removeTokenInfo } }) => ({
-      isLoggedIn,
-      isNeedSignUp,
-      isHydrated,
-      removeTokenInfo,
-    }),
-  );
+  const { mutate: mutateRefreshToken } = useRefreshTokenQuery();
+
+  // 초기 앱 진입 및 Foreground 복귀시마다 토큰 재발급 로직 수행여부 체크
+  useEffect(() => {
+    if (!isHydrated) {
+      return;
+    }
+
+    if (isNeedRefreshToken && tokenInfo.refresh?.token) {
+      mutateRefreshToken(
+        { refreshToken: tokenInfo.refresh.token },
+        {
+          onSuccess: ({ data }) => {
+            if (!data.atk || !data.rtk) {
+              return;
+            }
+            // 토큰 재발급이 완료 되었을 경우
+            setTokenInfo({ access: data.atk, refresh: data.rtk });
+            setIsLoggedIn(true);
+            setIsNeedRefreshToken(false);
+            setIsNeedSignUp(false);
+          },
+        },
+      );
+    }
+
+    const subscription = AppState.addEventListener('change', state => {
+      if (state === 'active') {
+        if (isNeedRefreshToken && tokenInfo.refresh?.token) {
+          mutateRefreshToken(
+            { refreshToken: tokenInfo.refresh.token },
+            {
+              onSuccess: ({ data }) => {
+                if (!data.atk || !data.rtk) {
+                  return;
+                }
+                // 토큰 재발급이 완료 되었을 경우
+                setTokenInfo({ access: data.atk, refresh: data.rtk });
+                setIsLoggedIn(true);
+                setIsNeedRefreshToken(false);
+                setIsNeedSignUp(false);
+              },
+            },
+          );
+        }
+      }
+    });
+
+    return () => subscription.remove();
+  }, [
+    tokenInfo,
+    isNeedRefreshToken,
+    isHydrated,
+    setIsLoggedIn,
+    setIsNeedRefreshToken,
+    setIsNeedSignUp,
+    setTokenInfo,
+    mutateRefreshToken,
+  ]);
 
   // useEffect(() => {
   //   removeTokenInfo();
